IMU only logging and USB data transfer
Two modules:

1) nocrlf: called from the terminal. Sets the USB port (/dev/ttyACM0) to
not expand linefeeds (LF, \n) into carriage-return-linefeeds (CRLF,
\r\n) using "termios". Can be issued even when the USB terminal is
already actively communicating.
The reason for this is to use the USB terminal to transfer data files
using the "cat" command, which prints raw bytes to the port, like this:
nsh> cat /fs/microsd/sess001/log001.bin
That prints all the characters in the file to the port. Assuming you log
the output on your desk computer, you will have a copy of the file you
requested.

2) sdlogimu: stripped down version of "sdlog2" suitable for logging only
IMU and barometer data, which is what I need to do.