Python Script - data transfer over usb
Python script to get a data file via the computer's USB connection to
the NSH terminal on the PX4. It issues the "nocrlf" command (and some
others), and intercepts all the non-data characters that come back
before and after.

Each user will have to edit the script though, for file names, etc. It's
really a template.